requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Security-Too==5.5.2
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-Caching==2.3.0
flask-cors==5.0.1
Flask-Excel==0.0.7
Flask-WTF==1.2.1
Flask-RESTful==0.3.10
SQLAlchemy==2.0.36
Werkzeug==3.0.4
WTForms==3.1.2
Jinja2==3.1.4
MarkupSafe==2.1.5
itsdangerous==2.2.0
blinker==1.8.2
passlib==1.7.4
email_validator==2.2.0
greenlet==3.1.1
cachelib==0.9.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
lml==0.1.0
texttable==1.7.0
chardet==5.2.0
gunicorn==22.0.0
